Add PhotoUpload and PhotoUploadQuestion models
For responses to photo upload questions, we want
to set the response’s value to the id of the
photo upload record to improve traceability.

We should be careful these don’t go out of sync so
add a data integrity check when the response’s
value is parsed.

Ideally, we’d do this as a validation but there’s
a chicken-and-egg problem since response depends
on the id of a photo upload and photo upload
depends on the id of a response.