Authenticate users via api tokens
The token is provided as the password of of the basic auth credentials.
If this matches a user, get their viewpoint (which now includes all
their user_roles and roles). Otherwise, respond with 401 and show the
username / password popup in a web browser.

If the user has the 'admin' role, they can access admin resources.