Present photos with responses
This is so that Attachments are created in the mobile app which are
pulled from the server if they don't exist on the device. Usually the
user's device will have the photos they uploaded, but it might not if
they're using the app across multiple devices at the same time.